+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## 1.1.1 (2024-02-11)
+
+### Notable Changes
+
+* Better error handling. No functionality changes.
+  * Check for required config
+  * Telnet connection issues
+
 ## 1.1.0 (2024-02-07)
 
 ### Notable Changes